Massive Revision to code, some edits have been discarded halfway through.
The original plan was to transform PlayerController in a master controller dedicated only to propagating events that listeners would subscribe to.

The purpose was to isolate and make some functionalities on the player controller strictly private, like health changing functionalities and so on. However, editing the behaviour of input risulted in a problematic outcome, which was not beneficial in any way whatsoever in terms of code implementation, since there already is a subscription to an event stream via GameInputAction Input Auto Generated Script.

There still is a possible use case for this but it might need further enhancement and elab to be handled correctly. The intention is to move the player action dictionary ion the PlayerActionController and use it to notify subscribers that a given action has been performed. It might be beneficial in some scenarios where actions need to be tracked, reacted to, but not implemented directly into that controller.